Fixed observers, this is a workaround for whenever SiteController#batch_page_status_refresh is called, (there usually won't be a user context present).

+++ app/models/page_observer.rb
@@ -2,16 +2,22 @@ class PageObserver < ActiveRecord::Observer
   cattr_accessor :current_user
 
   def before_create(page)
-    raise UserPagesExtension::AccessDenied unless current_user_can?(:create, page.parent)
+    if current_user
+      raise UserPagesExtension::AccessDenied unless current_user_can?(:create, page.parent)
+    end
   end
 
   def before_destroy(page)
-    raise UserPagesExtension::AccessDenied unless current_user_can?(:destroy, page)
-    page.permissions.destroy_all
+    if current_user
+      raise UserPagesExtension::AccessDenied unless current_user_can?(:destroy, page)
+      page.permissions.destroy_all
+    end
   end
 
   def before_update(page)
-    raise UserPagesExtension::AccessDenied unless current_user_can?(:update, page)
+    if current_user
+      raise UserPagesExtension::AccessDenied unless current_user_can?(:update, page)
+    end
   end
 
   def after_create(page)
